Revert "Merge branch 'dev' into create-dashboard-ui-and-chart-framework"
This reverts commit 43129728d4b09beabcd2c8241602a06a13c2b6cd, reversing
changes made to 584fc39aac03c704c360a60cc5ea0b9822f9d564.
