+++ client/src/App.js
@@ -3,11 +3,8 @@ import { MuiThemeProvider } from "@material-ui/core";
 import { BrowserRouter, Route } from "react-router-dom";
 
 import { theme } from "./themes/theme";
-import LandingPage from "./pages/Landing";
-import SignupPage from './pages/Signup';
-import LoginPage from './pages/Login';
 import Dashboard from "./pages/Dashboard";
-// import ButtonAppBar from "./pages/ButtonAppBar";
+import ButtonAppBar from "./pages/ButtonAppBar";
 import SideBar from "./pages/SideBar";
 import CreateReceipt from "./pages/CreateReceipt";
 
@@ -17,10 +14,8 @@ function App() {
   return (
     <MuiThemeProvider theme={theme}>
       <BrowserRouter>
-        <Route path="/signup" component={SignupPage} />
-        <Route path="/login" component={LoginPage} />
-        {/* <ButtonAppBar /> */}
-        {/* <SideBar /> */}
+        <ButtonAppBar />
+        <SideBar />
         <Route exact path="/" component={Dashboard} />
         <Route path="/receipt" component={CreateReceipt} />
       </BrowserRouter>
