+++ client/src/pages/ButtonAppBar.js
@@ -1,0 +1,38 @@
+import React from 'react';
+import { makeStyles } from '@material-ui/core/styles';
+import AppBar from '@material-ui/core/AppBar';
+import Toolbar from '@material-ui/core/Toolbar';
+import Typography from '@material-ui/core/Typography';
+import Button from '@material-ui/core/Button';
+import { Link } from 'react-router-dom';
+
+const useStyles = makeStyles(theme => ({
+  root: {
+    flexGrow: 1,
+  },
+  appBar: {
+    zIndex: theme.zIndex.drawer + 1,
+  },
+  title: {
+    flexGrow: 1,
+  },
+}));
+
+function ButtonAppBar() {
+  const classes = useStyles();
+
+  return (
+    <div className={classes.root}>
+      <AppBar position="fixed" className={classes.appBar}>
+        <Toolbar>
+          <Typography variant="h6" align="center" className={classes.title}>
+            Application Name
+          </Typography>
+          <Button component={ Link } to="/receipt" variant="outlined" color="inherit">Upload Receipt</Button>
+        </Toolbar>
+      </AppBar>
+    </div>
+  );
+}
+
+export default ButtonAppBar;
