+++ server/.env
@@ -1,10 +1,5 @@
 TEAM_NAME="Ethan, Isabel and William"
 FLASK_ENV="development"
-
 DB_USERNAME=""
 DB_PASSWORD=""
 DB_NAME=""
-
-S3_ACCESS_KEY=""
-S3_SECRET_KEY=""
-S3_BUCKET_NAME=""
