+++ server/Pipfile
@@ -8,12 +8,8 @@ verify_ssl = true
 [packages]
 flask = "*"
 flask-sqlalchemy = "*"
-flask-bcrypt = "*"
 psycopg2 = "*"
-pyjwt = "*"
-email-validator = "*"
 psycopg2-binary = "*"
-boto3 = "*"
 
 [requires]
 python_version = "3.7"
