+++ server/config.py
@@ -1,12 +1,6 @@
 import os
 
 TEAM_NAME = os.environ['TEAM_NAME']
-
 DB_USERNAME = os.environ['DB_USERNAME']
 DB_PASSWORD = os.environ['DB_PASSWORD']
 DB_NAME = os.environ['DB_NAME']
-
-S3_ACCESS_KEY = os.environ.get('S3_ACCESS_KEY')
-S3_SECRET_KEY = os.environ.get('S3_SECRET_KEY')
-S3_BUCKET_NAME = os.environ.get('S3_BUCKET_NAME')
-S3_LOCATION = 'http://{}.s3.amazonaws.com/'.format(BUCKET_NAME)
